requirements.txt
annotated-types==0.7.0
apprise==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
geographiclib==2.0
geopy==2.4.1
idna==3.10
Markdown==3.7
oauthlib==3.2.2
pydantic==2.10.3
pydantic-extra-types==2.10.1
pydantic_core==2.27.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
aiohttp==3.11.11
